pyproject.toml
[project]
name = "microblog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elasticsearch>=9.1.1",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-babel>=4.0.0",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-moment>=1.0.6",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "langdetect>=1.0.9",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "aiosmtpd>=1.4.6",
]
